# Reuse `svg` getter package template parts