--- a/scripts/build/templates/icon-object.js
+++ b/scripts/build/templates/icon-object.js
@@ -2,7 +2,7 @@
   title: '%s',
   slug: '%s',
   get svg() {
-    return '<svg role="img" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><title>%s</title><path d="' + this.path + '"/></svg>';
+    return a + '%s' + b + this.path + c;
   },
   path: '%s',
   source: '%s',