--- a/scripts/build/templates/index.js
+++ b/scripts/build/templates/index.js
@@ -1,5 +1,7 @@
 console.warn('Deprecation warning: The `simple-icons` entrypoint will be removed in the next major. Please switch to using `import * as icons from "simple-icons/icons"` if you need an object with all the icons.')
 
+%s
+
 var icons = {%s};
 
 Object.defineProperty(icons, "Get", {