The reduction in size of `icons.js` is not a lot but is something (in bytes): `4348959` -> `4158916`

### Basic performance test

Not very scientific, just median times of a few executions.

Before:

```
$ time node -e "for (let i=0; i<500000; i++) { require('./icons')['siElsevier'].svg }"

real	0m0,191s
user	0m0,186s
sys	0m0,016s
```

After is a bit slower:

```
real	0m0,198s
user	0m0,183s
sys	0m0,027s
```